requirements.txt
flask==3.1.1
flask-cors==6.0.1
selenium==4.34.2
webdriver-manager==4.0.2
fpdf==1.7.2
python-pptx==1.0.2
Pillow==11.3.0
